Unit tests for light and full DID passing
