+++ packages/did/src/DidDetailsv2/LightDidDetailsV2.spec.ts
@@ -5,233 +5,274 @@
  * found in the LICENSE file in the root directory of this source tree.
  */
 
-import { BN } from '@polkadot/util'
-import { randomAsHex } from '@polkadot/util-crypto'
+import { DidDocumentV2 } from '@kiltprotocol/types'
+import { Crypto } from '@kiltprotocol/utils'
+import { keypairToMultibaseKey, parse } from '../Did2.utils.js'
 
-import { ConfigService } from '@kiltprotocol/config'
-import type {
-  CryptoCallbacksV2,
-  DidDocumentV2,
-  KiltKeyringPair,
-  SubmittableExtrinsic,
-} from '@kiltprotocol/types'
-
-import {
-  ApiMocks,
-  TestUtilsV2
-} from '../../../../tests/testUtils'
-import { generateDidAuthenticatedTx } from '../Did2.chain.js'
 import * as Did from './index.js'
-
-const augmentedApi = ApiMocks.createAugmentedApi()
-const mockedApi: any = ApiMocks.getMockedApi()
-ConfigService.set({ api: mockedApi })
-
-jest.mock('../Did2.chain')
-jest
-  .mocked(generateDidAuthenticatedTx)
-  .mockResolvedValue({} as SubmittableExtrinsic)
+import type { NewService } from './DidDetailsV2.js'
 
 /*
+ * Functions tested:
+ * - createLightDidDocument
+ * - parseDocumentFromLightDid
+ *
  * Functions tested in integration tests:
  * - getKeysForExtrinsic
  * - authorizeExtrinsic
  */
 
-describe('When creating an instance from the chain', () => {
-  describe('authorizeBatch', () => {
-    let keypair: KiltKeyringPair
-    let sign: CryptoCallbacksV2.SignCallback
-    let fullDid: DidDocumentV2.DidDocument
+describe('When creating an instance from the details', () => {
+  it('correctly assign the right sr25519 authentication key, x25519 encryption key, and service endpoints', () => {
+    const authKey = Crypto.makeKeypairFromSeed(undefined, 'sr25519')
+    const encKey = Crypto.makeEncryptionKeypairFromSeed(
+      new Uint8Array(32).fill(1)
+    )
+    const service: NewService[] = [
+      {
+        id: '#service-1',
+        type: ['type-1'],
+        serviceEndpoint: ['x:url-1'],
+      },
+      {
+        id: '#service-2',
+        type: ['type-21', 'type-22'],
+        serviceEndpoint: ['x:url-21', 'x:url-22'],
+      },
+    ]
 
-    beforeAll(async () => {
-      const keyTool = TestUtilsV2.makeSigningKeyTool()
-      keypair = keyTool.keypair
-      fullDid = await TestUtilsV2.createLocalDemoFullDidFromKeypair(
-        keyTool.keypair
+    const lightDid = Did.createLightDidDocument({
+      authentication: [authKey],
+      keyAgreement: [encKey],
+      service,
+    })
+
+    expect(lightDid).toEqual(<DidDocumentV2.DidDocument>{
+      id: `did:kilt:light:00${authKey.address}:z17GNCdxLqMYTMC5pnnDrPZGxLEFcXvDamtGNXeNkfSaFf8cktX6erFJiQy8S3ugL981NNys7Rz8DJiaNPZi98v1oeFVL7PjUGNTz1g3jgZo4VgQri2SYHBifZFX9foHZH4DreZXFN66k5dPrvAtBpFXaiG2WZkkxsnxNWxYpqWPPcxvbTE6pJbXxWKjRUd7rog1h9vjA93QA9jMDxm6BSGJHACFgSPUU3UTLk2kjNwT2bjZVvihVFu1zibxwHjowb7N6UQfieJ7ny9HnaQy64qJvGqh4NNtpwkhwm5DTYUoAeAhjt3a6TWyxmBgbFdZF7`,
+      authentication: ['#authentication'],
+      keyAgreement: ['#encryption'],
+      verificationMethod: [
+        {
+          controller: `did:kilt:light:00${authKey.address}:z17GNCdxLqMYTMC5pnnDrPZGxLEFcXvDamtGNXeNkfSaFf8cktX6erFJiQy8S3ugL981NNys7Rz8DJiaNPZi98v1oeFVL7PjUGNTz1g3jgZo4VgQri2SYHBifZFX9foHZH4DreZXFN66k5dPrvAtBpFXaiG2WZkkxsnxNWxYpqWPPcxvbTE6pJbXxWKjRUd7rog1h9vjA93QA9jMDxm6BSGJHACFgSPUU3UTLk2kjNwT2bjZVvihVFu1zibxwHjowb7N6UQfieJ7ny9HnaQy64qJvGqh4NNtpwkhwm5DTYUoAeAhjt3a6TWyxmBgbFdZF7`,
+          id: '#authentication',
+          publicKeyMultibase: keypairToMultibaseKey({
+            publicKey: authKey.publicKey,
+            type: 'sr25519',
+          }),
+          type: 'MultiKey',
+        },
+        {
+          controller: `did:kilt:light:00${authKey.address}:z17GNCdxLqMYTMC5pnnDrPZGxLEFcXvDamtGNXeNkfSaFf8cktX6erFJiQy8S3ugL981NNys7Rz8DJiaNPZi98v1oeFVL7PjUGNTz1g3jgZo4VgQri2SYHBifZFX9foHZH4DreZXFN66k5dPrvAtBpFXaiG2WZkkxsnxNWxYpqWPPcxvbTE6pJbXxWKjRUd7rog1h9vjA93QA9jMDxm6BSGJHACFgSPUU3UTLk2kjNwT2bjZVvihVFu1zibxwHjowb7N6UQfieJ7ny9HnaQy64qJvGqh4NNtpwkhwm5DTYUoAeAhjt3a6TWyxmBgbFdZF7`,
+          id: '#encryption',
+          publicKeyMultibase: keypairToMultibaseKey({
+            publicKey: encKey.publicKey,
+            type: 'x25519',
+          }),
+          type: 'MultiKey',
+        },
+      ],
+      service: [
+        {
+          id: '#service-1',
+          type: ['type-1'],
+          serviceEndpoint: ['x:url-1'],
+        },
+        {
+          id: '#service-2',
+          type: ['type-21', 'type-22'],
+          serviceEndpoint: ['x:url-21', 'x:url-22'],
+        },
+      ],
+    })
+  })
+
+  it('correctly assign the right ed25519 authentication key and encryption key', () => {
+    const authKey = Crypto.makeKeypairFromSeed()
+    const encKey = Crypto.makeEncryptionKeypairFromSeed(
+      new Uint8Array(32).fill(1)
+    )
+
+    const lightDid = Did.createLightDidDocument({
+      authentication: [authKey],
+      keyAgreement: [encKey],
+    })
+
+    expect(parse(lightDid.id).address).toStrictEqual(authKey.address)
+
+    expect(lightDid).toEqual(<DidDocumentV2.DidDocument>{
+      id: `did:kilt:light:01${authKey.address}:z15dZSRuzEPTFnBErPxqJie4CmmQH1gYKSQYxmwW5Qhgz5Sr7EYJA3J65KoC5YbgF3NGoBsTY2v6zwj1uDnZzgXzLy8R72Fhjmp8ujY81y2AJc8uQ6s2pVbAMZ6bnvaZ3GVe8bMjY5MiKFySS27qRi`,
+      authentication: ['#authentication'],
+      keyAgreement: ['#encryption'],
+      verificationMethod: [
+        {
+          controller: `did:kilt:light:01${authKey.address}:z15dZSRuzEPTFnBErPxqJie4CmmQH1gYKSQYxmwW5Qhgz5Sr7EYJA3J65KoC5YbgF3NGoBsTY2v6zwj1uDnZzgXzLy8R72Fhjmp8ujY81y2AJc8uQ6s2pVbAMZ6bnvaZ3GVe8bMjY5MiKFySS27qRi`,
+          id: '#authentication',
+          publicKeyMultibase: keypairToMultibaseKey({
+            publicKey: authKey.publicKey,
+            type: 'ed25519',
+          }),
+          type: 'MultiKey',
+        },
+        {
+          controller: `did:kilt:light:01${authKey.address}:z15dZSRuzEPTFnBErPxqJie4CmmQH1gYKSQYxmwW5Qhgz5Sr7EYJA3J65KoC5YbgF3NGoBsTY2v6zwj1uDnZzgXzLy8R72Fhjmp8ujY81y2AJc8uQ6s2pVbAMZ6bnvaZ3GVe8bMjY5MiKFySS27qRi`,
+          id: '#encryption',
+          publicKeyMultibase: keypairToMultibaseKey({
+            publicKey: encKey.publicKey,
+            type: 'x25519',
+          }),
+          type: 'MultiKey',
+        },
+      ],
+    })
+  })
+
+  it('throws for unsupported authentication key type', () => {
+    const authKey = Crypto.makeKeypairFromSeed(undefined, 'ecdsa')
+    const invalidInput = {
+      // Not an authentication key type
+      authentication: [authKey],
+    }
+    expect(() =>
+      Did.createLightDidDocument(
+        invalidInput as unknown as Did.CreateDocumentInput
       )
-      sign = keyTool.getSignCallback(fullDid)
-    })
+    ).toThrowError()
+  })
 
-    describe('.addSingleTx()', () => {
-      it('fails if the extrinsic does not require a DID', async () => {
-        const extrinsic = augmentedApi.tx.indices.claim(1)
-        await expect(async () =>
-          Did.authorizeBatch({
-            did: fullDid.id,
-            batchFunction: augmentedApi.tx.utility.batchAll,
-            extrinsics: [extrinsic, extrinsic],
-            sign,
-            submitter: keypair.address,
-          })
-        ).rejects.toMatchInlineSnapshot(
-          '[DidBatchError: Can only batch extrinsics that require a DID signature]'
-        )
-      })
-
-      it('fails if the extrinsic is a utility (batch) extrinsic containing valid extrinsics', async () => {
-        const extrinsic = augmentedApi.tx.utility.batch([
-          await augmentedApi.tx.ctype.add('test-ctype'),
-        ])
-        const batchFunction =
-          jest.fn() as unknown as typeof mockedApi.tx.utility.batchAll
-        await Did.authorizeBatch({
-          did: fullDid.id,
-          batchFunction,
-          extrinsics: [extrinsic, extrinsic],
-          sign,
-          submitter: keypair.address,
-        })
-
-        expect(batchFunction).toHaveBeenCalledWith([extrinsic, extrinsic])
-      })
-
-      it('adds different batches requiring different keys', async () => {
-        const ctype1Extrinsic = await augmentedApi.tx.ctype.add(randomAsHex(32))
-        const ctype2Extrinsic = await augmentedApi.tx.ctype.add(randomAsHex(32))
-        const delegationExtrinsic1 =
-          await augmentedApi.tx.delegation.createHierarchy(
-            randomAsHex(32),
-            randomAsHex(32)
-          )
-        const delegationExtrinsic2 =
-          await augmentedApi.tx.delegation.createHierarchy(
-            randomAsHex(32),
-            randomAsHex(32)
-          )
-        const ctype3Extrinsic = await augmentedApi.tx.ctype.add(randomAsHex(32))
-        const ctype4Extrinsic = await augmentedApi.tx.ctype.add(randomAsHex(32))
-
-        const batchFunction =
-          jest.fn() as unknown as typeof augmentedApi.tx.utility.batchAll
-        const extrinsics = [
-          ctype1Extrinsic,
-          ctype2Extrinsic,
-          delegationExtrinsic1,
-          delegationExtrinsic2,
-          ctype3Extrinsic,
-          ctype4Extrinsic,
-        ]
-        await Did.authorizeBatch({
-          did: fullDid.id,
-          batchFunction,
-          extrinsics,
-          nonce: new BN(0),
-          sign,
-          submitter: keypair.address,
-        })
-
-        expect(batchFunction).toHaveBeenCalledWith([
-          ctype1Extrinsic,
-          ctype2Extrinsic,
-        ])
-        expect(batchFunction).toHaveBeenCalledWith([
-          delegationExtrinsic1,
-          delegationExtrinsic2,
-        ])
-        expect(batchFunction).toHaveBeenCalledWith([
-          ctype3Extrinsic,
-          ctype4Extrinsic,
-        ])
-      })
-    })
-
-    // TODO: complete these tests once SDK has been refactored to work with generic api object
-    describe('.build()', () => {
-      it('throws if batch is empty', async () => {
-        await expect(async () =>
-          Did.authorizeBatch({
-            did: fullDid.id,
-            batchFunction: augmentedApi.tx.utility.batchAll,
-            extrinsics: [],
-            sign,
-            submitter: keypair.address,
-          })
-        ).rejects.toMatchInlineSnapshot(
-          '[DidBatchError: Cannot build a batch with no transactions]'
-        )
-      })
-      it.todo('successfully create a batch with only 1 extrinsic')
-      it.todo('successfully create a batch with 1 extrinsic per required key')
-      it.todo('successfully create a batch with 2 extrinsics per required key')
-      it.todo(
-        'successfully create a batch with 1 extrinsic per required key, repeated two times'
+  it('throws for unsupported encryption key type', () => {
+    const authKey = Crypto.makeKeypairFromSeed()
+    const encKey = Crypto.makeEncryptionKeypairFromSeed()
+    const invalidInput = {
+      authentication: [authKey],
+      // Not an encryption key type
+      keyAgreement: [{ publicKey: encKey.publicKey, type: 'bls' }],
+    }
+    expect(() =>
+      Did.createLightDidDocument(
+        invalidInput as unknown as Did.CreateDocumentInput
       )
-    })
+    ).toThrowError()
   })
 })
 
-const mockApi = ApiMocks.createAugmentedApi()
+describe('When creating an instance from a URI', () => {
+  it('correctly assign the right authentication key, encryption key, and service endpoints', () => {
+    const authKey = Crypto.makeKeypairFromSeed(undefined, 'sr25519')
+    const encKey = Crypto.makeEncryptionKeypairFromSeed(
+      new Uint8Array(32).fill(1)
+    )
+    const endpoints: NewService[] = [
+      {
+        id: '#service-1',
+        type: ['type-1'],
+        serviceEndpoint: ['x:url-1'],
+      },
+      {
+        id: '#service-2',
+        type: ['type-21', 'type-22'],
+        serviceEndpoint: ['x:url-21', 'x:url-22'],
+      },
+    ]
+    // We are sure this is correct because of the described case above
+    const expectedLightDid = Did.createLightDidDocument({
+      authentication: [authKey],
+      keyAgreement: [encKey],
+      service: endpoints,
+    })
 
-describe('When creating an instance from the chain', () => {
-  it('Should return correct KeyRelationship for single valid call', () => {
-    const keyRelationship = Did.getVerificationMethodRelationshipForTx(
-      mockApi.tx.attestation.add(new Uint8Array(32), new Uint8Array(32), null)
-    )
-    expect(keyRelationship).toBe('assertionMethod')
+    const { address } = parse(expectedLightDid.id)
+    const builtLightDid = Did.parseDocumentFromLightDid(expectedLightDid.id)
+
+    expect(builtLightDid).toStrictEqual(expectedLightDid)
+    expect(builtLightDid).toStrictEqual(<DidDocumentV2.DidDocument>{
+      id: `did:kilt:light:00${address}:z17GNCdxLqMYTMC5pnnDrPZGxLEFcXvDamtGNXeNkfSaFf8cktX6erFJiQy8S3ugL981NNys7Rz8DJiaNPZi98v1oeFVL7PjUGNTz1g3jgZo4VgQri2SYHBifZFX9foHZH4DreZXFN66k5dPrvAtBpFXaiG2WZkkxsnxNWxYpqWPPcxvbTE6pJbXxWKjRUd7rog1h9vjA93QA9jMDxm6BSGJHACFgSPUU3UTLk2kjNwT2bjZVvihVFu1zibxwHjowb7N6UQfieJ7ny9HnaQy64qJvGqh4NNtpwkhwm5DTYUoAeAhjt3a6TWyxmBgbFdZF7`,
+      authentication: ['#authentication'],
+      keyAgreement: ['#encryption'],
+      verificationMethod: [
+        {
+          controller: `did:kilt:light:00${authKey.address}:z17GNCdxLqMYTMC5pnnDrPZGxLEFcXvDamtGNXeNkfSaFf8cktX6erFJiQy8S3ugL981NNys7Rz8DJiaNPZi98v1oeFVL7PjUGNTz1g3jgZo4VgQri2SYHBifZFX9foHZH4DreZXFN66k5dPrvAtBpFXaiG2WZkkxsnxNWxYpqWPPcxvbTE6pJbXxWKjRUd7rog1h9vjA93QA9jMDxm6BSGJHACFgSPUU3UTLk2kjNwT2bjZVvihVFu1zibxwHjowb7N6UQfieJ7ny9HnaQy64qJvGqh4NNtpwkhwm5DTYUoAeAhjt3a6TWyxmBgbFdZF7`,
+          id: '#authentication',
+          publicKeyMultibase: keypairToMultibaseKey({
+            publicKey: authKey.publicKey,
+            type: 'sr25519',
+          }),
+          type: 'MultiKey',
+        },
+        {
+          controller: `did:kilt:light:00${authKey.address}:z17GNCdxLqMYTMC5pnnDrPZGxLEFcXvDamtGNXeNkfSaFf8cktX6erFJiQy8S3ugL981NNys7Rz8DJiaNPZi98v1oeFVL7PjUGNTz1g3jgZo4VgQri2SYHBifZFX9foHZH4DreZXFN66k5dPrvAtBpFXaiG2WZkkxsnxNWxYpqWPPcxvbTE6pJbXxWKjRUd7rog1h9vjA93QA9jMDxm6BSGJHACFgSPUU3UTLk2kjNwT2bjZVvihVFu1zibxwHjowb7N6UQfieJ7ny9HnaQy64qJvGqh4NNtpwkhwm5DTYUoAeAhjt3a6TWyxmBgbFdZF7`,
+          id: '#encryption',
+          publicKeyMultibase: keypairToMultibaseKey({
+            publicKey: encKey.publicKey,
+            type: 'x25519',
+          }),
+          type: 'MultiKey',
+        },
+      ],
+      service: [
+        {
+          id: '#service-1',
+          type: ['type-1'],
+          serviceEndpoint: ['x:url-1'],
+        },
+        {
+          id: '#service-2',
+          type: ['type-21', 'type-22'],
+          serviceEndpoint: ['x:url-21', 'x:url-22'],
+        },
+      ],
+    })
   })
-  it('Should return correct KeyRelationship for batched call', () => {
-    const keyRelationship = Did.getVerificationMethodRelationshipForTx(
-      mockApi.tx.utility.batch([
-        mockApi.tx.attestation.add(
-          new Uint8Array(32),
-          new Uint8Array(32),
-          null
-        ),
-        mockApi.tx.attestation.add(
-          new Uint8Array(32),
-          new Uint8Array(32),
-          null
-        ),
-      ])
-    )
-    expect(keyRelationship).toBe('assertionMethod')
+
+  it('fail if a fragment is present according to the options', () => {
+    const authKey = Crypto.makeKeypairFromSeed()
+    const encKey = Crypto.makeEncryptionKeypairFromSeed()
+    const service: NewService[] = [
+      {
+        id: '#service-1',
+        type: ['type-1'],
+        serviceEndpoint: ['x:url-1'],
+      },
+      {
+        id: '#service-2',
+        type: ['type-21', 'type-22'],
+        serviceEndpoint: ['x:url-21', 'x:url-22'],
+      },
+    ]
+
+    // We are sure this is correct because of the described case above
+    const expectedLightDid = Did.createLightDidDocument({
+      authentication: [authKey],
+      keyAgreement: [encKey],
+      service,
+    })
+
+    const uriWithFragment: DidDocumentV2.DidUri = `${expectedLightDid.id}#authentication`
+
+    expect(() => Did.parseDocumentFromLightDid(uriWithFragment, true)).toThrow()
+    expect(() =>
+      Did.parseDocumentFromLightDid(uriWithFragment, false)
+    ).not.toThrow()
   })
-  it('Should return correct KeyRelationship for batchAll call', () => {
-    const keyRelationship = Did.getVerificationMethodRelationshipForTx(
-      mockApi.tx.utility.batchAll([
-        mockApi.tx.attestation.add(
-          new Uint8Array(32),
-          new Uint8Array(32),
-          null
-        ),
-        mockApi.tx.attestation.add(
-          new Uint8Array(32),
-          new Uint8Array(32),
-          null
-        ),
-      ])
-    )
-    expect(keyRelationship).toBe('assertionMethod')
-  })
-  it('Should return correct KeyRelationship for forceBatch call', () => {
-    const keyRelationship = Did.getVerificationMethodRelationshipForTx(
-      mockApi.tx.utility.forceBatch([
-        mockApi.tx.attestation.add(
-          new Uint8Array(32),
-          new Uint8Array(32),
-          null
-        ),
-        mockApi.tx.attestation.add(
-          new Uint8Array(32),
-          new Uint8Array(32),
-          null
-        ),
-      ])
-    )
-    expect(keyRelationship).toBe('assertionMethod')
-  })
-  it('Should return undefined for batch with mixed KeyRelationship calls', () => {
-    const keyRelationship = Did.getVerificationMethodRelationshipForTx(
-      mockApi.tx.utility.forceBatch([
-        mockApi.tx.attestation.add(
-          new Uint8Array(32),
-          new Uint8Array(32),
-          null
-        ),
-        mockApi.tx.web3Names.claim('awesomename'),
-      ])
-    )
-    expect(keyRelationship).toBeUndefined()
+
+  it('fail if the URI is not correct', () => {
+    const validKiltAddress = Crypto.makeKeypairFromSeed()
+    const incorrectURIs = [
+      'did:kilt:light:sdasdsadas',
+      // @ts-ignore not a valid DID uri
+      'random-uri',
+      'did:kilt:light',
+      'did:kilt:light:',
+      // Wrong auth key encoding
+      `did:kilt:light:11${validKiltAddress}`,
+      // Full DID
+      `did:kilt:${validKiltAddress}`,
+      // Random encoded details
+      `did:kilt:light:00${validKiltAddress}:randomdetails`,
+    ]
+    incorrectURIs.forEach((uri) => {
+      expect(() =>
+        Did.parseDocumentFromLightDid(uri as DidDocumentV2.DidUri)
+      ).toThrow()
+    })
   })
 })
